:wrench: Fixed addTo function taking args by adding a static cast from bool to uint8_t for the same reason as before. Bool not beeing correct in network.

+++ plugins/components/visible/Visible.hpp
@@ -111,7 +111,7 @@ public:
             throw std::runtime_error("Invalid number of arguments for Visible component");
         
         bool visibility = std::any_cast<bool>(args[0]);
-        engine.getRegistry().componentManager().addComponent<Components::Visible>(to, engine.newComponent<Components::Visible>(visibility));
+        engine.getRegistry().componentManager().addComponent<Components::Visible>(to, engine.newComponent<Components::Visible>(static_cast<uint8_t>(visibility)));
     }
 
     /**
